Arrays

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -10,7 +10,7 @@
  * Return: A pointer to the destination buffer @dest.
  */
 
-char *_memcpy(char *dest, char *src, size_t n)
+char *_memcpy(char *dest, char *src, unsigned int n)
 {
 	unsigned int index;
 	char *destination = dest;
